+++ src/actions/system.h
@@ -1,0 +1,12 @@
+/*
+ SPDX-License-Identifier: GPL-3.0-or-later
+ myGPIOd (c) 2020-2023 Juergen Mang <mail@jcgames.de>
+ https://github.com/jcorporation/mympd
+*/
+
+#ifndef MYGPIOD_ACTIONS_SYSTEM_H
+#define MYGPIOD_ACTIONS_SYSTEM_H
+
+void action_system(const char *cmd);
+
+#endif