+++ src/server/idle.h
@@ -9,8 +9,8 @@
 
 #include "src/lib/config.h"
 
-bool handle_idle(struct t_list_node *node);
-bool handle_noidle(struct t_config *config, struct t_list_node *node);
-bool send_idle_events(struct t_list_node *node);
+bool handle_idle(struct t_list_node *client_node);
+bool handle_noidle(struct t_config *config, struct t_list_node *client_node);
+bool send_idle_events(struct t_list_node *client_node);
 
 #endif