+++ src/server/protocol.h
@@ -11,16 +11,14 @@
 
 #include <stdbool.h>
 
-// Socket server default responses
-#define DEFAULT_ERROR_MSG_PREFIX "ERROR\nmessage:"
-#define DEFAULT_OK_MSG_PREFIX "OK\n"
-#define DEFAULT_END_MSG "END\n"
-
 #define CMDS(X) \
+    X(CMD_INVALID) \
     X(CMD_CLOSE) \
     X(CMD_IDLE) \
-    X(CMD_INVALID) \
     X(CMD_NOIDLE) \
+    X(CMD_GPIOLIST) \
+    X(CMD_GPIOGET) \
+    X(CMD_GPIOSET) \
     X(CMD_COUNT)
 
 /**
@@ -36,7 +34,12 @@ enum cmd_ids {
     CMDS(GEN_ENUM)
 };
 
-bool server_protocol_handler(struct t_config *config, struct t_list_node *node);
+struct t_cmd_options {
+    sds *args;
+    int len;
+};
+
+bool server_protocol_handler(struct t_config *config, struct t_list_node *client_node);
 enum cmd_ids get_cmd_id(const char *cmd);
 const char *get_cmd_name(enum cmd_ids cmd_id);