+++ src/server/response.c
@@ -12,40 +12,41 @@
  * It sends the response from buf_out to the client.
  * @param data pointer to client data
  */
-void server_response_start(struct t_client_data *data) {
-    sdsclear(data->buf_out);
+void server_response_start(struct t_client_data *client_data) {
+    sdsclear(client_data->buf_out);
 }
 
 /**
  * Appends the message to the output buffer
- * @param data pointer to client data
- * @param message message to append
+ * @param client_data pointer to client data
+ * @param fmt printf style format string
+ * @param ... variadic arguments for the format string
  */
-void server_response_append(struct t_client_data *data, const char *fmt, ...) {
+void server_response_append(struct t_client_data *client_data, const char *fmt, ...) {
     va_list args;
     va_start(args, fmt);
-    data->buf_out = sdscatvprintf(data->buf_out, fmt, args);
+    client_data->buf_out = sdscatvprintf(client_data->buf_out, fmt, args);
     va_end(args);
 }
 
 /**
  * Sets the client state to writing.
  * It sends the response from buf_out to the client.
- * @param data pointer to client data
+ * @param client_data pointer to client data
  */
-void server_response_end(struct t_client_data *data) {
-    data->state = CLIENT_SOCKET_STATE_WRITING;
-    data->bytes_out = 0;
-    data->events = POLLOUT;
+void server_response_end(struct t_client_data *client_data) {
+    client_data->state = CLIENT_SOCKET_STATE_WRITING;
+    client_data->bytes_out = 0;
+    client_data->events = POLLOUT;
 }
 
 /**
  * Shortcut for server_response_start, server_response_append and server_response_end.
- * @param data pointer to client data
+ * @param client_data pointer to client data
  * @param message message to send
  */
-void server_response_send(struct t_client_data *data, const char *message) {
-    server_response_start(data);
-    server_response_append(data, "%s", message);
-    server_response_end(data);
+void server_response_send(struct t_client_data *client_data, const char *message) {
+    server_response_start(client_data);
+    server_response_append(client_data, "%s", message);
+    server_response_end(client_data);
 }