+++ src/server/response.h
@@ -9,10 +9,15 @@
 
 #include "src/server/socket.h"
 
-void server_response_send(struct t_client_data *data, const char *message);
-void server_response_start(struct t_client_data *data);
-void server_response_append(struct t_client_data *data, const char *fmt, ...)
+// Socket server default responses
+#define DEFAULT_ERROR_MSG_PREFIX "ERROR\nmessage:"
+#define DEFAULT_OK_MSG_PREFIX "OK\n"
+#define DEFAULT_END_MSG "END\n"
+
+void server_response_send(struct t_client_data *client_data, const char *message);
+void server_response_start(struct t_client_data *client_data);
+void server_response_append(struct t_client_data *client_data, const char *fmt, ...)
     __attribute__ ((format (printf, 2, 3)));
-void server_response_end(struct t_client_data *data);
+void server_response_end(struct t_client_data *client_data);
 
 #endif